requirements.txt
Keras==2.4.3
numpy==1.19.5
Flask==1.1.2
tensorflow-cpu==2.4.1
gunicorn==20.0.4
Werkzeug==0.16.0
Jinja2==2.10.1
itsdangerous==1.1.0
MarkupSafe==1.1.1
PyJWT==1.7.1
h5py==2.10.0
